refactor(repository): key validation moved to static class
